add quantized test

+++ Source/Falcor/Utils/Neural/MLPCuda.cpp
@@ -130,21 +130,21 @@ void MLPCuda::inferFp16(int* packedInput, float* output, int width, int height, 
     );
 }
 
-void MLPCuda::inferInt8Test(float* testInput, float* output, int width, int height, float scale)
+void MLPCuda::inferInt8Test(int* testInput, float* output, int width, int height, float scale)
 {
     launchInferInt8TexTest((int*)mpInt8Buffer->getGpuAddress(), testInput, mHTexObj, mDTexObj, mUTexObj, output, width, height, scale);
 }
 
-void MLPCuda::inferFp32Test(float* testInput, float* output, int width, int height, float scale)
+void MLPCuda::inferFp32Test(int* testInput, float* output, int width, int height, float scale)
 {
     launchInferFp32TexTest((float*)mpFp32Buffer->getGpuAddress(), testInput, mHTexObj, mDTexObj, mUTexObj, output, width, height, scale);
 }
 
-void MLPCuda::inferFp16Test(float* testInput, float* output, int width, int height, float scale)
+void MLPCuda::inferFp16Test(int* testInput, float* output, int width, int height, float scale)
 {
     launchInferFp16TexTest((__half*)mpFp16Buffer->getGpuAddress(), testInput, mHTexObj, mDTexObj, mUTexObj, output, width, height, scale);
 }
-void MLPCuda::inferInt8ACFTest(float* testInput, float* output, int width, int height, float scale)
+void MLPCuda::inferInt8ACFTest(int* testInput, float* output, int width, int height, float scale)
 {
     launchInferInt8TexACFTest(
         (int*)mpInt8Buffer->getGpuAddress(),
