+++ Source/Falcor/Utils/Neural/MLPCuda.h
@@ -27,10 +27,10 @@ public:
     void inferFp16(int* packedInput, float* output, int width, int height, int* valid, float scale);
 
     // for infer speed test
-    void inferInt8Test(float* testInput, float* output, int width, int height, float scale);
-    void inferFp32Test(float* testInput, float* output, int width, int height, float scale);
-    void inferFp16Test(float* testInput, float* output, int width, int height, float scale);
-    void inferInt8ACFTest(float* testInput, float* output, int width, int height, float scale);
+    void inferInt8Test(int* testInput, float* output, int width, int height, float scale);
+    void inferFp32Test(int* testInput, float* output, int width, int height, float scale);
+    void inferFp16Test(int* testInput, float* output, int width, int height, float scale);
+    void inferInt8ACFTest(int* testInput, float* output, int width, int height, float scale);
 
     ref<Buffer> mpInt8Buffer;
     ref<Buffer> mpSampleBuffer;
